remove FileWatching from compilation engine to fix DTH memory leak
