+++ src/Microsoft.Dnx.Compilation/FileSystem/FileWatcher.cs
@@ -204,37 +204,4 @@ namespace Microsoft.Dnx.Compilation.FileSystem
             return false;
         }
     }
-
-    public sealed class NoopWatcher : IFileWatcher
-    {
-        public static readonly NoopWatcher Instance = new NoopWatcher();
-
-        private NoopWatcher()
-        {
-        }
-
-        public bool WatchFile(string path)
-        {
-            return true;
-        }
-
-        // Suppressing warning CS0067: The event is never used
-#pragma warning disable 0067
-
-        public event Action<string> OnChanged;
-
-#pragma warning restore 0067
-
-        public void WatchDirectory(string path, string extension)
-        {
-        }
-
-        public void Dispose()
-        {
-        }
-
-        public void WatchProject(string path)
-        {
-        }
-    }
 }
