+++ src/Microsoft.Dnx.Runtime/Compilation/CompilationEngineContext.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Runtime.Versioning;
+using Microsoft.Dnx.Compilation;
 using Microsoft.Dnx.Runtime.Common.DependencyInjection;
 
 namespace Microsoft.Dnx.Runtime.Compilation
@@ -10,14 +11,16 @@ namespace Microsoft.Dnx.Runtime.Compilation
 
         public LibraryManager LibraryManager { get; }
         public IProjectGraphProvider ProjectGraphProvider { get; }
+        public IFileWatcher FileWatcher { get; }
         public FrameworkName TargetFramework { get; }
         public string Configuration { get; }
         public IServiceProvider Services { get { return _services; } }
 
-        public CompilationEngineContext(LibraryManager libraryManager, IProjectGraphProvider projectGraphProvider, IServiceProvider services, FrameworkName targetFramework, string configuration)
+        public CompilationEngineContext(LibraryManager libraryManager, IProjectGraphProvider projectGraphProvider, IFileWatcher fileWatcher, IServiceProvider services, FrameworkName targetFramework, string configuration)
         {
             LibraryManager = libraryManager;
             ProjectGraphProvider = projectGraphProvider;
+            FileWatcher = fileWatcher;
             TargetFramework = targetFramework;
             Configuration = configuration;
